+++ EchoMMO-Backend/src/main/java/com/echommo/service/InventoryService.java
@@ -1,5 +1,3 @@
-// File: EchoMMO-Backend/src/main/java/com/echommo/service/InventoryService.java
-
 package com.echommo.service;
 
 import com.echommo.entity.User;
@@ -9,14 +7,14 @@ import java.util.List;
 public interface InventoryService {
     List<UserItem> getInventory(Integer charId);
 
-    // [FIX] Đổi Long -> Integer để chiều lòng client
-    void equipItem(Integer charId, Integer userItemId);
-    void unequipItem(Integer charId, Integer userItemId);
-    UserItem enhanceItem(Integer charId, Integer userItemId);
+    // [FIXED] Long userItemId
+    void equipItem(Integer charId, Long userItemId);
+    void unequipItem(Integer charId, Long userItemId);
+    UserItem enhanceItem(Integer charId, Long userItemId);
 
     User expandInventory(User user);
     void addItemToInventory(User user, Integer itemId, int quantity);
 
-    // [FIX] Đổi Long -> Integer
-    UserItem repairItem(User user, Integer userItemId);
+    // [FIXED] Long userItemId
+    UserItem repairItem(User user, Long userItemId);
 }
